Return single contig for each cell based on filtering and sorting

Return single contig for each cell based on filtering and sorting

canonicalize_by_subset(ccdb, ..., tie_break_keys = c("umis", "reads"),
  contig_fields = tie_break_keys, order = 1)

Arguments

ccdb
`ContigCellDB`

...
unquoted expressions passed to `dplyr::filter` that will be applied to the `contig_tbl`

tie_break_keys
columns used to sort the `contig_tbl` in **decreasing** order

contig_fields
columns in the `contig_tbl` that will be copied over to the `cell_tbl`

order
`integer` specifying which entry (ordinal) will be selected from the sorted contig_tbl, within each cell

Value

`ContigCellDB` with additional fields in `cell_tbl`

Examples

# Report beta chain with highest umi-count, breaking ties with reads
beta = canonicalize_by_subset(ccdb_ex, chain == 'TRB',
tie_break_keys = c('umis', 'reads'),
contig_fields = c('umis', 'reads', 'chain', 'v_gene', 'd_gene', 'j_gene'))
head(beta$cell_tbl)

#> # A tibble: 6 x 9
#> # Groups:   pop, sample, barcode [6]
#>    umis reads chain v_gene d_gene j_gene  pop   sample barcode           
#>   <dbl> <dbl> <chr> <chr>  <chr>  <chr>   <chr> <chr>  <chr>             
#> 1     6 37898 TRB   TRBV31 None   TRBJ1-5 b6    4      AAAGTAGTCGCGCCAA-1
#> 2    NA    NA NA    NA     NA     NA      b6    4      AACCATGCATTTGCCC-1
#> 3     6 33548 TRB   TRBV5  TRBD2  TRBJ2-2 b6    4      AACTGGTGTCTGATCA-1
#> 4     6 26928 TRB   TRBV4  TRBD2  TRBJ2-4 b6    4      AAGCCGCAGTAAGTAC-1
#> 5     4 18435 TRB   TRBV1  None   TRBJ2-4 b6    4      AAGTCTGGTTCAACCA-1
#> 6     9 46156 TRB   TRBV5  TRBD2  TRBJ2-7 b6    4      ACACCAAAGTCCAGGA-1

# Only adds fields to `cell_tbl`
stopifnot(all.equal(beta$cell_tbl[ccdb_ex$cell_pk],
ccdb_ex$cell_tbl[ccdb_ex$cell_pk]))
#Report cdr3 with highest UMI count, but only when > 5 UMIs support it
umi5 = canonicalize_by_subset(ccdb_ex, umis > 5,
tie_break_keys = c('umis', 'reads'), contig_fields = c('umis', 'cdr3'))
stopifnot(all(umi5$cell_tbl$umis > 5, na.rm = TRUE))