Cluster "And" intersection

For each contig present in both X and Y, a new cluster is defined that combines cluster identities in both X and Y. In the resulting ContigCellDB, two contigs are in the same cluster if they are in the same cluster in X and the same cluster in Y. X and Y must have matching contig_pk. The contig_tbl has fields from X for contigs present in both X and Y. The cell_tbl from X is carried forward unchanged, while the cluster_tbl in the result contains the mapping between the ancestral clustering, and the derived.

cland(X, Y, new_pk)

Arguments

X

ContigCellDB

Y

ContigCellDB

new_pk

optional character naming the new pk.

Examples

data(ccdb_ex)
ccdb_germ = cluster_germline(ccdb_ex, cluster_pk = 'germline_idx')
ccdb_cdr3 = cdhit_ccdb(ccdb_ex, 'cdr3_nt', type = 'DNA',
cluster_name = 'DNA97', identity = .965, min_length = 12, G = 1)
ccdb_cdr3 = cland(ccdb_cdr3, ccdb_germ)