Tests for independence between clusters and labels using permutation

This tests a statistic for association between clusters and labels by permuting the link between the two Both clusters and labels should be factors. Each observation represents a cell. For example, clusters could be a clonal ID, while labels is a subject ID. statistic is any

cluster_permute_test(clusters, labels, covariates, statistic,
  n_perm = 1000, alternative = c("two.sided", "less", "greater"), ...)

Arguments

clusters
factor of length n

labels
factor of length n

covariates
optional factor of length n.

statistic
function of clusters, labels and optionally covariates that returns a numeric of length 1.

n_perm
number of permutations.

alternative
`character` naming the direction `statistic` should be fall under the alternative hypothesis

...
passed to statistic

Value

a list containing the observed value of the statistic, its expectation (under independence), a p-value, and the monte carlo standard error (of the expected value).

Examples

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following object is masked from ‘package:testthat’:
#> 
#>     matches

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

purity = function(clusters, labels, covariates){
n_label_cluster = data_frame(labels = labels, clusters = clusters) %>%
    group_by(clusters, labels) %>% summarize(n = n()) %>% ungroup()
singletons = mean(n_label_cluster$n == 1)
singletons
}

clusters = c(1, 1, 1, 2, 2, 3, 3)
labels = c('A', 'A', 'B', 'B', 'B', 'C', 'C')
covariates = c('X', 'X', 'Y', 'Y', 'Y', 'Y', 'Y')
cluster_permute_test(clusters, labels, statistic = purity, n_perm  = 50)

#> Warning: `data_frame()` is deprecated, use `tibble()`.
#> This warning is displayed once per session.

#> $observed
#> [1] 0.25
#> 
#> $expected
#> [1] 0.8173333
#> 
#> $p.value
#> [1] 0.02
#> 
#> $mc.se
#> [1] 0.0210655
#>