Filtered and annotated contigs of TCR from mice

Data for c57bl6 and balbc mice TCR were downloaded from 10x Genomics website as shown in system.file('script/10XMouseTCR_v3_chem.R', package = 'CellaRepertorium'). Additional processing of these data is done in the vignette mouse_tcell_qc and are serialized to serve as an examples for other vignettes and documentation.

data(contigs_qc)

Format

A data frame of 3399 contigs and 22 fields, all except 4 are originally defined in https://support.10xgenomics.com/single-cell-vdj/software/pipelines/latest/output/annotation#contig The following fields were defined ex post facto.

anno_file: Path to original csv file

pop: Mouse strain.

sample: An artificial "replicate" from the original data defined by subsampling with replacement

celltype: The putative cell type of the contig.