Tests for independence between labels and covariates using permutation of cells

This tests a statistic for association between labels (for instance, cluster/clonal ID) and covariates (for instance, subject or treatment) by permuting the link between the two. Each observation represents a cell. statistic is any function of labels

cluster_permute_test(
  ccdb,
  cell_covariate_keys,
  cell_label_key = ccdb$cluster_pk,
  cell_stratify_keys,
  statistic,
  n_perm,
  alternative = c("two.sided", "less", "greater"),
  sanity_check_strata = TRUE,
  ...
)

Arguments

ccdb
ContigCellDB

cell_covariate_keys
character naming fields in ccdb$cell_tbl

cell_label_key
character naming a single field in ccdb$cell_tbl

cell_stratify_keys
optional character naming fields in ccdb$cell_tbl under which permutations of cell_label_key will occur. This means that the test will occur conditional on these covariates. Must be disjoint from cell_covariate_keys.

statistic
function of label (vector) and covariate (data.frame). Must return a scalar

n_perm
number of permutations to run

alternative
character naming the direction statistic should be fall under the alternative hypothesis

sanity_check_strata
logical, should cell_stratify_keys be checked for sanity?

...
passed to statistic

Value

a list containing the observed value of the statistic, its expectation (under independence), a p-value, and the Monte Carlo standard error (of the expected value).

See also

Examples

library(dplyr)
# covariate should name one or more columns in `cell_tbl`

cluster_idx = c(1, 1, 1, 2, 2, 3, 3)
subject = c('A', 'A', 'B', 'B', 'B', 'C', 'C')
contig_tbl = tibble(contig_pk = seq_along(cluster_idx), cluster_idx, subject)
ccdb_test = ContigCellDB(contig_tbl = contig_tbl, contig_pk = 'contig_pk',
cell_pk = c('contig_pk', 'subject', 'cluster_idx'), cluster_pk = 'cluster_idx')
ccdb_test$cell_tbl

#> # A tibble: 7 x 3
#>   contig_pk subject cluster_idx
#>       <int> <chr>         <dbl>
#> 1         1 A                 1
#> 2         2 A                 1
#> 3         3 B                 1
#> 4         4 B                 2
#> 5         5 B                 2
#> 6         6 C                 3
#> 7         7 C                 3

purity_ex = function(cluster_idx, subject) {
n_label_cluster = tibble(cluster_idx = cluster_idx, subject = subject) %>%
group_by(cluster_idx, subject) %>% summarize(n = dplyr::n()) %>% ungroup()
#Average number of singleton clusters for each subject
singletons = mean(n_label_cluster$n == 1)
singletons
}
cluster_permute_test(ccdb_test, 'subject', 'cluster_idx',
statistic = purity, n_perm  = 50)

#> $observed
#> [1] 0.25
#> 
#> $expected
#> [1] 0.8173333
#> 
#> $p.value
#> [1] 0.02
#> 
#> $mc.se
#> [1] 0.0210655
#>