Take the intersection of keys in tables in x

The cells in cell_tbl, and clusters in cluster_tbl can potentially be a superset of the contig_tbl.

equalize_ccdb(x, cell = TRUE, contig = TRUE, cluster = TRUE)

Arguments

x
ContigCellDB()

cell
logical equalize cells

contig
logical equalize contigs

cluster
logical equalize clusters

Value

ContigCellDB()

Details

equalize_ccdb(x, cell = TRUE) trims cells that aren't in contig_tbl or cluster_tbl.

equalize_ccdb(x, cluster = TRUE) trims clusters that aren't in contig_tbl.

equalize_ccdb(x, contig = TRUE) trims contigs that aren't cell_tbl or cluster_tbl.