Perform a dplyr::left_join() but check for non-key overlapping fields

Perform a dplyr join, but either warn if the two tables share non-key fields If overwrite = TRUE, then shared columns will pull from x otherwise a suffix will be added to y. To perform this check, by must be specified, and it is an error if it is not.

right_join_warn(...)

left_join_warn(x, y, by, overwrite = FALSE, join = left_join, ...)

Arguments

...
passed to joining function

x
A pair of data frames, data frame extensions (e.g. a tibble), or lazy data frames (e.g. from dbplyr or dtplyr). See Methods, below, for more details.

y
A pair of data frames, data frame extensions (e.g. a tibble), or lazy data frames (e.g. from dbplyr or dtplyr). See Methods, below, for more details.

by
character specifying columns in x and y to key on.

overwrite
logical -- should non-key fields in y be overwritten using x, or should a suffix (".y") be added

join
function giving the type of join to perform, eg, left, right, inner, outer.

Value

data.frame or tibble

Functions

right_join_warn: perform a dplyr::right_join()

Examples

left_join_warn(mtcars, mtcars, by  = 'mpg')

#> Warning: Tables share fields cyl, disp, hp, drat, wt, qsec, vs, am, gear, carb, which will gain a suffix in mtcars

#>     mpg cyl  disp  hp drat    wt  qsec vs am gear carb cyl.y disp.y hp.y drat.y
#> 1  21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4     6  160.0  110   3.90
#> 2  21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4     6  160.0  110   3.90
#> 3  21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4     6  160.0  110   3.90
#> 4  21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4     6  160.0  110   3.90
#> 5  22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1     4  108.0   93   3.85
#> 6  22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1     4  140.8   95   3.92
#> 7  21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1     6  258.0  110   3.08
#> 8  21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1     4  121.0  109   4.11
#> 9  18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2     8  360.0  175   3.15
#> 10 18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1     6  225.0  105   2.76
#> 11 14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4     8  360.0  245   3.21
#> 12 24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2     4  146.7   62   3.69
#> 13 22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2     4  108.0   93   3.85
#> 14 22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2     4  140.8   95   3.92
#> 15 19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4     6  167.6  123   3.92
#> 16 19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4     8  400.0  175   3.08
#> 17 17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4     6  167.6  123   3.92
#> 18 16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3     8  275.8  180   3.07
#> 19 17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3     8  275.8  180   3.07
#> 20 15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3     8  275.8  180   3.07
#> 21 15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3     8  304.0  150   3.15
#> 22 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4     8  472.0  205   2.93
#> 23 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4     8  460.0  215   3.00
#> 24 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4     8  472.0  205   2.93
#> 25 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4     8  460.0  215   3.00
#> 26 14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4     8  440.0  230   3.23
#> 27 32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1     4   78.7   66   4.08
#> 28 30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2     4   75.7   52   4.93
#> 29 30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2     4   95.1  113   3.77
#> 30 33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1     4   71.1   65   4.22
#> 31 21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1     4  120.1   97   3.70
#> 32 15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2     8  318.0  150   2.76
#> 33 15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2     8  275.8  180   3.07
#> 34 15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2     8  304.0  150   3.15
#> 35 13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4     8  350.0  245   3.73
#> 36 19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2     6  167.6  123   3.92
#> 37 19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2     8  400.0  175   3.08
#> 38 27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1     4   79.0   66   4.08
#> 39 26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2     4  120.3   91   4.43
#> 40 30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2     4   75.7   52   4.93
#> 41 30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2     4   95.1  113   3.77
#> 42 15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4     8  351.0  264   4.22
#> 43 19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6     6  145.0  175   3.62
#> 44 15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8     8  301.0  335   3.54
#> 45 21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2     6  258.0  110   3.08
#> 46 21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2     4  121.0  109   4.11
#>     wt.y qsec.y vs.y am.y gear.y carb.y
#> 1  2.620  16.46    0    1      4      4
#> 2  2.875  17.02    0    1      4      4
#> 3  2.620  16.46    0    1      4      4
#> 4  2.875  17.02    0    1      4      4
#> 5  2.320  18.61    1    1      4      1
#> 6  3.150  22.90    1    0      4      2
#> 7  3.215  19.44    1    0      3      1
#> 8  2.780  18.60    1    1      4      2
#> 9  3.440  17.02    0    0      3      2
#> 10 3.460  20.22    1    0      3      1
#> 11 3.570  15.84    0    0      3      4
#> 12 3.190  20.00    1    0      4      2
#> 13 2.320  18.61    1    1      4      1
#> 14 3.150  22.90    1    0      4      2
#> 15 3.440  18.30    1    0      4      4
#> 16 3.845  17.05    0    0      3      2
#> 17 3.440  18.90    1    0      4      4
#> 18 4.070  17.40    0    0      3      3
#> 19 3.730  17.60    0    0      3      3
#> 20 3.780  18.00    0    0      3      3
#> 21 3.435  17.30    0    0      3      2
#> 22 5.250  17.98    0    0      3      4
#> 23 5.424  17.82    0    0      3      4
#> 24 5.250  17.98    0    0      3      4
#> 25 5.424  17.82    0    0      3      4
#> 26 5.345  17.42    0    0      3      4
#> 27 2.200  19.47    1    1      4      1
#> 28 1.615  18.52    1    1      4      2
#> 29 1.513  16.90    1    1      5      2
#> 30 1.835  19.90    1    1      4      1
#> 31 2.465  20.01    1    0      3      1
#> 32 3.520  16.87    0    0      3      2
#> 33 3.780  18.00    0    0      3      3
#> 34 3.435  17.30    0    0      3      2
#> 35 3.840  15.41    0    0      3      4
#> 36 3.440  18.30    1    0      4      4
#> 37 3.845  17.05    0    0      3      2
#> 38 1.935  18.90    1    1      4      1
#> 39 2.140  16.70    0    1      5      2
#> 40 1.615  18.52    1    1      4      2
#> 41 1.513  16.90    1    1      5      2
#> 42 3.170  14.50    0    1      5      4
#> 43 2.770  15.50    0    1      5      6
#> 44 3.570  14.60    0    1      5      8
#> 45 3.215  19.44    1    0      3      1
#> 46 2.780  18.60    1    1      4      2

left_join_warn(mtcars, mtcars, by = 'mpg', overwrite = TRUE)

#> Warning: Overwriting fields cyl, disp, hp, drat, wt, qsec, vs, am, gear, carb in table mtcars

#>     mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> 1  21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> 2  21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> 3  21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> 4  21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> 5  22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> 6  22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> 7  21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> 8  21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> 9  18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> 10 18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> 11 14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> 12 24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> 13 22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> 14 22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> 15 19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> 16 19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> 17 17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> 18 16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> 19 17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> 20 15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> 21 15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> 22 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> 23 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> 24 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> 25 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> 26 14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> 27 32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> 28 30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> 29 30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> 30 33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> 31 21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> 32 15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> 33 15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> 34 15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> 35 13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> 36 19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> 37 19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> 38 27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> 39 26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> 40 30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> 41 30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> 42 15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> 43 19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> 44 15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> 45 21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#> 46 21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2