Generate a list of tables representing clusters paired in cells

A contingency table of every combination of cluster_idx up to table_order is generated. Combinations that are found in at least min_expansion number of cells are reported. All cells that have these combinations are returned, as well as cells that only have orphan_level of matching cluster_idx.

pairing_tables(ccdb, canonicalize_fun = canonicalize_by_chain,
  table_order = 2, min_expansion = 2, orphan_level = 1,
  cluster_keys = character(), cluster_whitelist = NULL,
  cluster_blacklist = NULL)

Arguments

ccdb
ContigCellDB

canonicalize_fun
a function with signature canonicalize_fun(ContigCellDB, order = i) that for each cell, returns a single contig that depends on the order. For instance canonicalize_by_prevalence or canonicalize_by_chain.

table_order
Integer larger than 1. What order of cluster_idx will be paired, eg, order = 2 means that the most common and second most common cluster_idx will be sought for each cell

min_expansion
the minimal number of times a pairing needs to occur for it to be reported

orphan_level
Integer larger than 0 and less than or equal to table_order. Given that at least min_expansion cells are found that have table_order chains identical, how many cluster_idx pairs will we match on to select other cells. Example: ophan_level=1 means that cells that share just a single chain with the

cluster_keys
optional character naming additional columns in ccdb$cluster_tbl to be reported in the pairing

cluster_whitelist
a table of pairings or clusters that should always be reported. Here the clusters must be named "cluster_idx.1", "cluster_idx.2" (if order-2 pairs are being selected) rather than with `ccdb$cluster_pk``

cluster_blacklist
a table of pairings or clusters that will never be reported. Must be named as per cluster_whitelist.

Value

list of tables. The cell_tbl is keyed by the cell_identifiers, with fields "cluster_idx.1", "cluster_idx.2", etc, IDing the contigs present in each cell. "cluster_idx.1_fct" and "cluster_idx.2_fct" cast these fields to factors and are reordered to maximize the number of pairs along the diagonal. The idx1_tbl and idx2_tbl report information (passed in about the cluster_idx by feature_tbl.) The cluster_pair_tbl reports all pairings found of contigs, and the number of times observed.

Details

For example, if table_order=2 and min_expansion=2 then heavy/light or alpha/beta pairs found two or more times will be returned (as well as alpha-alpha pairs, etc, if those are present). If orphan_level=1 then all cells that share just a single chain with an expanded clone will be returned.

The cluster_idx.1_fct and cluster_idx.2_fct fields in cell_tbl, idx1_tbl, idx2_tbl are cast to factors and ordered such that pairings will tend to occur along the diagonal when they are cross-tabulated. This facilitates plotting.

Caveats and warnings

The cell_idx -> cluster_idx map is generally one-to-many, and is resolved by canonicalize_fun. For table_order>1, few collisions are expected as most cells will contain no more than 2 clusters. Any collisions are resolved by returning the most prevalent cluster, across samples. When two clusters are tied for most prevalent within a cell, the cluster_idx returned is arbitrary. Therefore, when table_order=1, it is strongly recommended to subset the cluster_tbl to just a single chain.

See also

Examples

library(dplyr)
tbl = tibble(clust_idx = gl(3, 2), cell_idx = rep(1:3, times = 2), contig_idx = 1:6)
ccdb = ContigCellDB(tbl, contig_pk = c('cell_idx', 'contig_idx'),
cell_pk = 'cell_idx', cluster_pk = 'clust_idx')
# no pairs found twice
pt1 = pairing_tables(ccdb, canonicalize_by_prevalence)

#> Warning: No pairs found

# all pairs found, found once.
pt2 = pairing_tables(ccdb, canonicalize_by_prevalence, min_expansion = 1)

#> Warning: Column `clust_idx` joining factors with different levels, coercing to character vector

pt2$cell_tbl

#> # A tibble: 3 x 6
#>   cell_idx cluster_idx.1 cluster_idx.2 max_pairs cluster_idx.1_f…
#>      <int> <fct>         <fct>             <int> <fct>           
#> 1        1 2             1                     1 2               
#> 2        2 3             1                     1 3               
#> 3        3 3             2                     1 3               
#> # … with 1 more variable: cluster_idx.2_fct <fct>

tbl2 = bind_rows(tbl, tbl %>% mutate(cell_idx = rep(4:6, times = 2)))
ccdb2 = ContigCellDB(tbl2, contig_pk = c('cell_idx', 'contig_idx'), cell_pk = 'cell_idx',
cluster_pk = 'clust_idx')
#all pairs found twice
pt3 = pairing_tables(ccdb2, canonicalize_by_prevalence, min_expansion = 1)

#> Warning: Column `clust_idx` joining factors with different levels, coercing to character vector

pt3$cell_tbl

#> # A tibble: 6 x 6
#>   cell_idx cluster_idx.1 cluster_idx.2 max_pairs cluster_idx.1_f…
#>      <int> <fct>         <fct>             <int> <fct>           
#> 1        1 2             1                     2 2               
#> 2        2 3             1                     2 3               
#> 3        3 3             2                     2 3               
#> 4        4 2             1                     2 2               
#> 5        5 3             1                     2 3               
#> 6        6 3             2                     2 3               
#> # … with 1 more variable: cluster_idx.2_fct <fct>

# `canonicalize_by_chain` expects fields `umis`, `reads`
# to break ties,  wrap the function to change this
ccdb2$contig_tbl = ccdb2$contig_tbl %>%
    mutate(umis = 1, reads = 1, chain = rep(c('TRA', 'TRB'), times = 6))
pt4 = pairing_tables(ccdb2, canonicalize_by_chain, min_expansion = 1, table_order = 2)